misc/busybot/app/requirements.txt
discord
python-dotenv
cowpy
